move envutil back out, gofmt
