+++ pkg/internal/envutil/envutil_test.go
@@ -20,7 +20,7 @@ import (
 	"testing"
 
 	v1alpha1 "github.com/google/kf/pkg/apis/kf/v1alpha1"
-	"github.com/google/kf/pkg/kf/internal/utils/envutil"
+	"github.com/google/kf/pkg/internal/envutil"
 	"github.com/google/kf/pkg/kf/testutil"
 	corev1 "k8s.io/api/core/v1"
 )
