+++ pkg/kf/apps/kfapp.go
@@ -17,7 +17,7 @@ package apps
 import (
 	v1alpha1 "github.com/google/kf/pkg/apis/kf/v1alpha1"
 
-	"github.com/google/kf/pkg/kf/internal/utils/envutil"
+	"github.com/google/kf/pkg/internal/envutil"
 	"github.com/google/kf/pkg/kf/sources"
 	corev1 "k8s.io/api/core/v1"
 	v1 "k8s.io/api/core/v1"
