+++ pkg/kf/cfutil/vcapapp.go
@@ -16,7 +16,7 @@ package cfutil
 
 import (
 	v1alpha1 "github.com/google/kf/pkg/apis/kf/v1alpha1"
-	"github.com/google/kf/pkg/kf/internal/utils/envutil"
+	"github.com/google/kf/pkg/internal/envutil"
 	corev1 "k8s.io/api/core/v1"
 )
