+++ pkg/kf/commands/apps/unset_env_test.go
@@ -22,7 +22,7 @@ import (
 	utils "github.com/google/kf/pkg/kf/internal/utils/cli"
 
 	"github.com/golang/mock/gomock"
-	"github.com/google/kf/pkg/kf/internal/utils/envutil"
+	"github.com/google/kf/pkg/internal/envutil"
 	"github.com/google/kf/pkg/kf/apps"
 	"github.com/google/kf/pkg/kf/apps/fake"
 	"github.com/google/kf/pkg/kf/commands/config"
